+++ civis/futures.py
@@ -9,12 +9,13 @@ import datetime
 import logging
 import time
 import threading
+import warnings
 
 import six
 
 from civis import APIClient
-from civis.base import DONE
-from civis.polling import PollableResult, _ResultPollingThread
+from civis.base import CivisAPIError, DONE
+from civis.polling import PollableResult
 
 from pubnub.pubnub import PubNub
 from pubnub.pnconfiguration import PNConfiguration, PNReconnectionPolicy
@@ -119,9 +120,15 @@ class CivisFuture(PollableResult):
                          client=client,
                          poll_on_creation=poll_on_creation)
 
-        if hasattr(client, 'channels'):
-            config, channels = self._pubnub_config()
-            self._pubnub = self._subscribe(config, channels)
+    def _begin_tracking(self, start_thread=False):
+        # Be sure to subscribe to the PubNub channel before polling.
+        # Otherwise the job might complete after polling and before
+        # we subscribe, causing us to miss the notification.
+        with self._condition:
+            if hasattr(self.client, 'channels') and not self.subscribed:
+                config, channels = self._pubnub_config()
+                self._pubnub = self._subscribe(config, channels)
+            super()._begin_tracking(start_thread)  # superclass does polling
 
     @property
     def subscribed(self):
@@ -134,13 +141,31 @@ class CivisFuture(PollableResult):
             if hasattr(self, '_pubnub'):
                 self._pubnub.unsubscribe_all()
 
+                # Pubnub doesn't close its open session, so do that ourselves
+                # to free up sockets. We have to access a private attribute,
+                # but this exists at least in (4.0.0 <= versions <= 4.0.11).
+                # After closing the Session, remove it so that PubNub
+                # can't reopen it.
+                self._pubnub._request_handler.session.close()
+                del self._pubnub._request_handler.session
+
+                # The PubNub object is no longer usable.
+                # Note that it can't be garbage collected because of circular
+                # references, so this represents a (small) memory leak.
+                del self._pubnub
+
     def _subscribe(self, pnconfig, channels):
         listener = JobCompleteListener(self._check_message,
                                        self._poll_and_set_api_result,
                                        self._reset_polling_thread)
         pubnub = PubNub(pnconfig)
         pubnub.add_listener(listener)
-        pubnub.subscribe().channels(channels).execute()
+
+        # Start our subscription 30 seconds in the past to catch any
+        # missed messages.
+        # https://www.pubnub.com/docs/python/api-reference-misc#time
+        token = int((time.time() - 30) * 10**7)
+        pubnub.subscribe().channels(channels).with_timetoken(token).execute()
         return pubnub
 
     def _pubnub_config(self):
@@ -151,7 +176,7 @@ class CivisFuture(PollableResult):
         pnconfig.cipher_key = channel_config['cipher_key']
         pnconfig.auth_key = channel_config['auth_key']
         pnconfig.ssl = True
-        pnconfig.reconnect_policy = PNReconnectionPolicy.LINEAR
+        pnconfig.reconnect_policy = PNReconnectionPolicy.EXPONENTIAL
         return pnconfig, channels
 
     def _check_message(self, message):
@@ -251,13 +276,8 @@ class ContainerFuture(CivisFuture):
                 # you shut down the old thread too soon after starting it.
                 # In practice this only happens when testing retries
                 # with extremely short polling intervals.
-                self._polling_thread = _ResultPollingThread(
-                    self._check_result, (), self.polling_interval)
-                self._polling_thread.start()
+                self._begin_tracking(start_thread=True)
 
-                if hasattr(self, '_pubnub'):
-                    # Subscribe to the new run's notifications endpoint
-                    self._pubnub = self._subscribe(*self._pubnub_config())
                 log.debug('Job ID %d / Run ID %d failed. Retrying '
                           'with run %d. %d retries remaining.',
                           self.job_id, orig_run_id,
@@ -279,7 +299,18 @@ class ContainerFuture(CivisFuture):
             elif not self.done():
                 # Cancel the job and store the result of the cancellation in
                 # the "finished result" attribute, `_result`.
-                self._result = self.client.scripts.post_cancel(self.job_id)
+                try:
+                    self._result = self.client.scripts.post_cancel(self.job_id)
+                except CivisAPIError as exc:
+                    if exc.status_code == 404:
+                        # The most likely way to get this error
+                        # is for the job to already be completed.
+                        return False
+                    else:
+                        warnings.warn("Unexpected error when attempting to "
+                                      "cancel job ID %d / run ID %d:\n%s" %
+                                      (self.job_id, self.run_id, str(exc)))
+                        return False
                 for waiter in self._waiters:
                     waiter.add_cancelled(self)
                 self._condition.notify_all()