+++ civis/tests/mocks.py
@@ -26,17 +26,17 @@ def create_client_mock(cache=TEST_SPEC_ALL):
         A `Mock` object which looks like an APIClient and which will
         error if any method calls have non-existent / misspelled parameters
     """
-    # Create a client from the cache. We'll use this for
-    # auto-speccing. Prevent it from trying to talk to the real API.
-    with mock.patch('requests.Session', mock.MagicMock):
-        real_client = APIClient(local_api_spec=cache, api_key='none',
-                                resources='all')
+    # Create a client from the cache. We'll use this for auto-speccing.
+    real_client = APIClient(local_api_spec=cache, api_key='none',
+                            resources='all')
     real_client._feature_flags = {'noflag': None}
     if hasattr(real_client, 'channels'):
         # Deleting "channels" causes the client to fall back on
         # regular polling for completion, which greatly eases testing.
         delattr(real_client, 'channels')
 
-    mock_client = mock.create_autospec(real_client, spec_set=True)
+    # Prevent the client from trying to talk to the real API when autospeccing
+    with mock.patch('requests.Session', mock.MagicMock):
+        mock_client = mock.create_autospec(real_client, spec_set=True)
 
     return mock_client